requirement.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==22.6.0
bs4==0.0.1
captcha-solver==0.1.5
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.0
colorama==0.4.6
cssselect2==0.4.1
distlib==0.3.7
dj-database-url==0.5.0
Django==4.2.3
django-crispy-forms==1.9.2
django-formtools==2.4.1
django-sms==0.6.0
djangorestframework==3.14.0
doctor==3.13.7
filelock==3.12.2
gunicorn==20.1.0
hospital==0.9
idna==2.10
isodate==0.6.1
jsonschema==2.6.0
lxml==4.9.3
MarkupSafe==2.0.0
mypy-extensions==1.0.0
numpy==1.25.1
pandas==2.0.3
pathspec==0.11.1
patient==0.0.1
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.6
pyairmore==0.2.1a1
pyarrow==12.0.1
PyJWT==2.1.0
pylance==0.5.5
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2020.1
PyYAML==3.13
reportlab==4.0.4
requests==2.25.1
rfc3987==1.3.8
simplejson==3.19.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.1
strict-rfc3339==0.7
svglib==1.1.0
tinycss2==1.1.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.4
virtualenv==20.24.2
webencodings==0.5.1
whitenoise==6.2.0
